Add get_state_would_return predicate
This predicate allows us to talk about what get_state(_aux) would yield without needing
to expose the definition of get_state.

Some lemmas are added to prove this predicate after set_state, to propagate it through
the trace (somewhat poorly so far), and to use it to draw conclusions about the result
of get_state.